MAINT: move ground state solver to capp module
