+++ smol/capp/__init__.py
@@ -1,0 +1,4 @@
+"""Cluster Applications module.
+
+Classes and functions implementing various tools based on CE, MC and related methods.
+"""
