+++ smol/capp/generate/__init__.py
@@ -1,0 +1,1 @@
+"""General tools for structure and occupancy generation."""
