+++ smol/capp/generate/solver/__init__.py
