+++ smol/capp/generate/solver/upper_bound/objectives.py
@@ -6,12 +6,14 @@ from typing import List, Tuple, Union
 import numpy as np
 from numpy.typing import ArrayLike
 
+from smol.capp.generate.solver.upper_bound.utils.indices import (
+    map_ewald_indices_to_variable_indices,
+)
 from smol.moca.processor import (
     ClusterDecompositionProcessor,
     ClusterExpansionProcessor,
     EwaldProcessor,
 )
-from smol.solver.upper_bound.utils.indices import map_ewald_indices_to_variable_indices
 
 __author__ = "Fengyu Xie"
