+++ smol/capp/generate/solver/upper_bound/utils/indices.py
