+++ smol/capp/generate/solver/upper_bound/utils/terms.py
