+++ smol/capp/generate/solver/upper_bound/variables.py
