+++ tests/test_solver/test_constraints.py
@@ -4,13 +4,13 @@ import itertools
 import pytest
 from pymatgen.core import Species
 
-from smol.moca.utils.occu import get_dim_ids_table, occu_to_counts
-from smol.solver.upper_bound.constraints import (
+from smol.capp.generate.solver import get_upper_bound_variables_from_sublattices
+from smol.capp.generate.solver.upper_bound.constraints import (
     get_upper_bound_composition_space_constraints,
     get_upper_bound_fixed_composition_constraints,
     get_upper_bound_normalization_constraints,
 )
-from smol.solver.upper_bound.variables import get_upper_bound_variables_from_sublattices
+from smol.moca.utils.occu import get_dim_ids_table, occu_to_counts
 
 from .utils import get_random_neutral_variable_values, get_random_variable_values
