+++ tests/test_solver/test_solver.py
@@ -7,11 +7,11 @@ import pytest
 from pymatgen.analysis.structure_matcher import StructureMatcher
 from pymatgen.core import Lattice, Structure
 
+from smol.capp.generate.solver import get_variable_values_from_occupancy
+from smol.capp.generate.solver.upper_bound.solver import UpperboundSolver
 from smol.cofe import ClusterExpansion, ClusterSubspace
 from smol.moca import Ensemble
 from smol.moca.utils.occu import get_dim_ids_table, occu_to_counts
-from smol.solver.upper_bound.solver import UpperboundSolver
-from smol.solver.upper_bound.variables import get_variable_values_from_occupancy
 
 from ..utils import assert_msonable
